client/requirements.txt
grpcio
grpcio-tools
opencv-python
numpy
onnxruntime
onnx